requirements.txt
accelerate==0.31.0
deepspeed==0.14.4
jsonlines==4.0.0
mpi4py_mpich==3.1.5
pandas==2.2.2
Pillow==10.0.1
pyarrow==16.1.0
xformers==0.0.23
python-dotenv==1.0.1
numpy==1.26.4
